+++ libs/fleXdLoggerClient/src/FleXdLogger.cpp
@@ -55,7 +55,7 @@ namespace {
                 return flexd::logger::MsgType::Enum::ERROR;
         }
     }
-    
+
     flexd::logger::LogLevel::Enum msgTypeToLogLevel(flexd::logger::MsgType::Enum msgType) {
         switch (msgType) {
             case flexd::logger::MsgType::Enum::VERBOSE:
@@ -112,20 +112,19 @@ namespace flexd {
         }
 
         void FleXdLogger::writeLog(LogLevel::Enum logLevel, std::time_t time, const std::string& level, const std::string& stream) {
-	  if(msgTypeToLogLevel(m_IPCClient->getLogLvlFilter()) <= logLevel){
-            writeLogToBuffer(logLevel, stream, time);
-            if (m_IPCClient && !m_IPCClient->isConnected()) {
-               std::cout << "FleXdLogger::[" << m_IPCClient->getName() << "][" << m_IPCClient->getAppID() << "][" << time <<"][" << level << "] : " << stream << std::endl;
+            if(msgTypeToLogLevel(m_IPCClient->getLogLvlFilter()) <= logLevel){
+                writeLogToBuffer(logLevel, stream, time);
+                if (m_IPCClient && !m_IPCClient->isConnected()) {
+                std::cout << "FleXdLogger::[" << m_IPCClient->getName() << "][" << m_IPCClient->getAppID() << "][" << time <<"][" << level << "] : " << stream << std::endl;
+                }
             }
-	  }
         }
 
         void FleXdLogger::writeLogToBuffer(const LogLevel::Enum logLevel, const std::string& stream, time_t time) {
-                LogData logData{time, logLevelToMsgType(logLevel), m_msgCount, stream};
-                m_logBuffer->push(std::move(logData));
+                m_logBuffer->push(LogData(time, logLevelToMsgType(logLevel), m_msgCount, stream));
                 m_msgCount++;
                 if (m_IPCClient) m_IPCClient->flushBuffer();
-            
+
         }
 
         void FleXdLogger::handshake() {